backend/pyproject.toml
[project]
name = "auto-research"
version = "0.1.0"
description = "Automated research pipeline using Claude CLI"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "pyjwt[crypto]>=2.8",
    "requests>=2.31",
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
